requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
python-dotenv==0.15.0
requests==2.25.1
Unidecode==1.1.2


